pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "aiosqlite>=0.22.1",
    "fastapi>=0.135.1",
    "mcp[cli]>=1.26.0",
    "pydantic-settings>=2.13.1",
    "python-dotenv>=1.2.2",
    "python-jose[cryptography]>=3.5.0",
    "sqlalchemy>=2.0.48",
    "uvicorn[standard]>=0.41.0",
]
